requirements.txt
flask
flask-session
python-dotenv
requests
azure-identity
tiktoken
langchain
langchain-openai
openai 